Consolidate ConfigService protocol
