+++ AuthenticatorShared/UI/DebugMenu/DebugMenuProcessorTests.swift
@@ -1,4 +1,5 @@
 import BitwardenKit
+import BitwardenKitMocks
 import XCTest
 
 @testable import AuthenticatorShared
