+++ AuthenticatorShared/UI/Vault/VaultItem/AuthenticatorKeyCapture/ManualEntryProcessorTests.swift
@@ -1,3 +1,4 @@
+import BitwardenKitMocks
 import XCTest
 
 @testable import AuthenticatorShared
