+++ BitwardenShared/UI/Auth/IntroCarousel/IntroCarouselProcessorTests.swift
@@ -1,3 +1,4 @@
+import BitwardenKitMocks
 import XCTest
 
 @testable import BitwardenShared
