+++ BitwardenShared/UI/Platform/Application/Utilities/CoordinatorTests.swift
@@ -1,3 +1,4 @@
+import BitwardenKitMocks
 import TestHelpers
 import XCTest
