+++ BitwardenShared/UI/Platform/DebugMenu/DebugMenuCoordinatorTests.swift
@@ -1,3 +1,4 @@
+import BitwardenKitMocks
 import SwiftUI
 import XCTest
